+++ spec/lib/stellar/client_spec.rb
@@ -32,7 +32,7 @@ describe Stellar::Client do
     it "returns the current details for the account", vcr: { record: :once, match_requests_on: [:method]} do
       response = client.account_info(account)
 
-      expect(response.id).to eq "GCQSESW66AX4ZRZB7QWCIXSPX2BD7KLOYSS33IUGDCLO4XCPURZEEC6R"
+      expect(response.id).to eq CONFIG[:source_address]
       expect(response.paging_token).to be_empty
       expect(response.sequence).to eq "346973227974715"
       expect(response.subentry_count).to eq 0
@@ -40,10 +40,10 @@ describe Stellar::Client do
       expect(response.flags).to include("auth_required" => false, "auth_revocable" => false)
       expect(response.balances).to include("balance" => "3494.9997500", "asset_type" => "native")
       expect(response.signers).to include(
-        "public_key" => "GCQSESW66AX4ZRZB7QWCIXSPX2BD7KLOYSS33IUGDCLO4XCPURZEEC6R",
+        "public_key" => CONFIG[:source_address],
         "weight" => 1,
         "type" => "ed25519_public_key",
-        "key"=>"GCQSESW66AX4ZRZB7QWCIXSPX2BD7KLOYSS33IUGDCLO4XCPURZEEC6R"
+        "key" => CONFIG[:source_address]
       )
       expect(response.data).to be_empty
     end
@@ -81,18 +81,25 @@ describe Stellar::Client do
     end
 
     context "alphanum4 asset" do
-      let(:destination) { Stellar::Account.from_seed(CONFIG[:destination_seed]) }
+      let(:issuer) { Stellar::Account.from_seed(CONFIG[:source_seed]) }
+      let(:destination) { Stellar::Account.random }
 
-      it "sends a alphanum4 asset to the destination", vcr: {record: :once, match_requests_on: [:method]} do
-        destination_info = client.account_info(destination)
-        old_balances = destination_info.balances
-        old_btc_balance = old_balances.find do |b|
-          b["asset_code"] == "BTC"
-        end["balance"].to_f
+      it("sends a alphanum4 asset to the destination", {
+        vcr: {record: :once, match_requests_on: [:method]},
+      }) do
+        client.create_account(
+          funder: issuer,
+          account: destination,
+          starting_balance: 2,
+        )
+
+        client.change_trust(
+          asset: [:alphanum4, "BTC", issuer.keypair],
+          source: destination,
+        )
 
         asset = Stellar::Asset.alphanum4("BTC", source.keypair)
         amount = Stellar::Amount.new(150, asset)
-
         client.send_payment(
           from: source,
           to: destination,
@@ -100,24 +107,31 @@ describe Stellar::Client do
         )
 
         destination_info = client.account_info(destination)
-        new_balances = destination_info.balances
-        new_btc_balance = new_balances.find do |b|
+        btc_balance = destination_info.balances.find do |b|
           b["asset_code"] == "BTC"
         end["balance"].to_f
 
-        expect(new_btc_balance - old_btc_balance).to eq 150.0
+        expect(btc_balance).to eq 150.0
       end
     end
 
     context "alphanum12 asset" do
-      let(:destination) { Stellar::Account.from_seed(CONFIG[:destination_seed]) }
+      let(:issuer) { Stellar::Account.from_seed(CONFIG[:source_seed]) }
+      let(:destination) { Stellar::Account.random }
 
-      it "sends a alphanum12 asset to the destination", vcr: {record: :once, match_requests_on: [:method]} do
-        destination_info = client.account_info(destination)
-        old_balances = destination_info.balances
-        old_btc_balance = old_balances.find do |b|
-          b["asset_code"] == "LONGNAME"
-        end["balance"].to_f
+      it("sends a alphanum12 asset to the destination", {
+        vcr: {record: :once, match_requests_on: [:method]},
+      }) do
+        client.create_account(
+          funder: issuer,
+          account: destination,
+          starting_balance: 2,
+        )
+
+        client.change_trust(
+          asset: [:alphanum12, "LONGNAME", issuer.keypair],
+          source: destination,
+        )
 
         asset = Stellar::Asset.alphanum12("LONGNAME", source.keypair)
         amount = Stellar::Amount.new(150, asset)
@@ -129,12 +143,11 @@ describe Stellar::Client do
         )
 
         destination_info = client.account_info(destination)
-        new_balances = destination_info.balances
-        new_btc_balance = new_balances.find do |b|
+        btc_balance = destination_info.balances.find do |b|
           b["asset_code"] == "LONGNAME"
         end["balance"].to_f
 
-        expect(new_btc_balance - old_btc_balance).to eq 150.0
+        expect(btc_balance).to eq 150.0
       end
     end
   end
@@ -166,6 +179,64 @@ describe Stellar::Client do
       it "accepts a cursor to return less data", vcr: {record: :once, match_requests_on: [:method]} do
         response = client.transactions(cursor: cursor)
         expect(response).to be_a(Stellar::TransactionPage)
+      end
+    end
+  end
+
+  describe "#change_trust" do
+    context "given an asset described as an array" do
+      let(:issuer) { Stellar::Account.from_seed(CONFIG[:source_seed]) }
+      let(:truster) { Stellar::Account.random }
+
+      it("creates, updates, or deletes a trustline", {
+        vcr: {record: :once, match_requests_on: [:method]},
+      }) do
+        client.create_account(
+          funder: issuer,
+          account: truster,
+          starting_balance: 2,
+        )
+
+        # Create trustline
+        client.change_trust(
+          asset: [:alphanum4, "BTC", issuer.keypair],
+          source: truster,
+        )
+
+        truster_info = client.account_info(truster)
+        btc_balance = truster_info.balances.find do |b|
+          b["asset_code"] == "BTC" && b["asset_issuer"] == issuer.address
+        end
+
+        expect(btc_balance).to_not be_nil
+
+        # Update trustline
+        client.change_trust(
+          asset: [:alphanum4, "BTC", issuer.keypair],
+          source: truster,
+          limit: 100,
+        )
+
+        truster_info = client.account_info(truster)
+        btc_balance = truster_info.balances.find do |b|
+          b["asset_code"] == "BTC" && b["asset_issuer"] == issuer.address
+        end
+
+        expect(btc_balance["limit"].to_f).to eq 100
+
+        # Delete trustline
+        client.change_trust(
+          asset: [:alphanum4, "BTC", issuer.keypair],
+          source: truster,
+          limit: 0,
+        )
+
+        truster_info = client.account_info(truster)
+        btc_balance = truster_info.balances.find do |b|
+          b["asset_code"] == "BTC" && b["asset_issuer"] == issuer.address
+        end
+
+        expect(btc_balance).to be_nil
       end
     end
   end